Часть функционала контроллера вынес в родительский класс Controller, от которого будут наследовать остальные контроллеры
