+++ controllers/CartController.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace app\controllers;
+
+use app\models\Cart;
+
+class CartController extends Controller {
+
+}
